# openapi/mvp.yaml
openapi: 3.1.0
info:
  title: 1Router MVP API
  version: 0.1.0
  description: Minimal contract for MVP routing and policy/key management workflows.
servers:
  - url: http://localhost:3000
paths:
...
  /v1/embeddings:
    post:
      summary: Create embeddings (stub)
      operationId: createEmbeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbeddingsRequest"
            example:
              model: openai/text-embedding-3-small
              input:
                - hello
                - world
      responses:
        "200":
          description: Embeddings created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsResponse"
              example:
                object: list
                data:
                  - object: embedding
                    embedding: [0.532, 0.05, 0.804124]
                    index: 0
                  - object: embedding
                    embedding: [0.552, 0.05, 0.845361]
                    index: 1
                model: openai/text-embedding-3-small
                usage:
                  prompt_tokens: 3
                  total_tokens: 3
                x_router:
                  provider: openai
                  stub: true
        "400":
          description: Invalid request or unsupported model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /v1/chat/completions:
    post:
      summary: Create chat completion (supports streaming stub)
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionsRequest"
            example:
              model: openai/gpt-4.1-mini
              messages:
                - role: user
                  content: Say hi
              stream: false
              temperature: 0.7
      responses:
        "200":
          description: Chat completion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionsResponse"
            text/event-stream:
              schema:
                type: string
        "400":
          description: Invalid request or unsupported model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /v1/responses:
    post:
      summary: Create response (non-streaming stub)
      operationId: createResponse
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponsesRequest"
            example:
              model: openai/gpt-4.1-mini
              input: Say hi
              stream: false
      responses:
        "200":
          description: Response created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponsesResponse"
        "400":
          description: Invalid request or unsupported model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
        "409":
          description: Idempotency key conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /v1/routing/decision/explain:
    post:
      summary: Explain routing decision factors (non-streaming)
      operationId: explainRoutingDecision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutingExplainRequest"
            example:
              model: router/auto
              routing_preset: success
              region_preference: EU
      responses:
        "200":
          description: Routing decision trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingExplainResponse"
        "400":
          description: Invalid request or unsupported model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /api/orgs/{orgId}/usage:
    get:
      summary: Read aggregated usage for an org
      operationId: getOrgUsage
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: group_by
          required: false
          schema:
            type: string
            enum: [hour, model]
      responses:
        "200":
          description: Usage report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUsageResponse"
              example:
                data:
                  org_id: org_demo
                  from: "2026-02-26T10:00:00.000Z"
                  to: "2026-02-26T12:00:00.000Z"
                  group_by: hour
                  provisional: false
                  finalized_at: "2026-02-26T12:00:00.000Z"
                  totals:
                    bucket: total
                    requests: 28
                    input_tokens: 4200
                    output_tokens: 1600
                    total_tokens: 5800
                    cost_usd: 0.231
                    platform_fee_usd: 0.0231
                    provisional: true
                    finalized_at: null
                  summary:
                    provisional: true
                    finalized_at: null
                    totals:
                      bucket: total
                      requests: 28
                      input_tokens: 4200
                      output_tokens: 1600
                      total_tokens: 5800
                      cost_usd: 0.231
                      platform_fee_usd: 0.0231
                      provisional: true
                      finalized_at: null
                  buckets:
                    - bucket: "2026-02-26T10:00:00.000Z"
                      requests: 13
                      input_tokens: 1900
                      output_tokens: 700
                      total_tokens: 2600
                      cost_usd: 0.102
                      platform_fee_usd: 0.0102
                      provisional: true
                      finalized_at: null
                    - bucket: "2026-02-26T11:00:00.000Z"
                      requests: 15
                      input_tokens: 2300
                      output_tokens: 900
                      total_tokens: 3200
                      cost_usd: 0.129
                      platform_fee_usd: 0.0129
                      provisional: false
                      finalized_at: "2026-02-26T12:00:00.000Z"
                meta:
                  request_id: req_usage_001
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /api/orgs/{orgId}/invoice:
    get:
      summary: Read monthly invoice line items for an org
      operationId: getOrgInvoice
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: "^[0-9]{4}-[0-9]{2}$"
      responses:
        "200":
          description: Monthly invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgInvoiceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /api/keys:
    post:
      summary: Create provider API key (returns plaintext once)
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
            example:
              provider: openai
              label: prod-openai
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
    get:
      summary: List configured provider keys
      operationId: listApiKeys
      responses:
        "200":
          description: API keys list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeysListResponse"
              example:
                data:
                  - id: key_01
                    provider: openai
                    label: prod-openai
                    last4: sk12
                    status: active
                  - id: key_02
                    provider: anthropic
                    label: staging-anthropic
                    last4: abcd
                    status: inactive
                meta:
                  request_id: req_keys_001
  /api/orgs:
    get:
      summary: List organizations (MVP control-plane)
      operationId: listOrgs
      responses:
        "200":
          description: Organizations list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgsListResponse"
              example:
                data:
                  - id: org_01
                    name: Acme Inc
                    status: active
                    created_at: "2026-01-01T00:00:00.000Z"
                    updated_at: "2026-01-01T00:00:00.000Z"
                meta:
                  request_id: req_orgs_001
    post:
      summary: Create organization (MVP)
      operationId: createOrg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
            example:
              name: Acme Inc
              status: active
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgRecordResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /api/orgs/{orgId}:
    patch:
      summary: Update organization (MVP)
      operationId: updateOrg
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgRequest"
            example:
              status: inactive
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgRecordResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
  /api/projects:
    get:
      summary: List projects (MVP control-plane)
      operationId: listProjects
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectsListResponse"
              example:
                data:
                  - id: prj_01
                    org_id: org_01
                    name: Acme Default
                    status: active
                    created_at: "2026-01-01T00:00:00.000Z"
                    updated_at: "2026-01-01T00:00:00.000Z"
                meta:
                  request_id: req_projects_001
    post:
      summary: Create project (MVP)
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
            example:
              org_id: org_01
              name: Acme Default
              status: active
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRecordResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingsError"
...
components:
  schemas:
...
    EmbeddingsRequest:
      type: object
      additionalProperties: false
      required: [model, input]
      properties:
        model:
          type: string
          minLength: 1
        input:
          oneOf:
            - type: string
              minLength: 1
            - type: array
              minItems: 1
              items:
                type: string
                minLength: 1
...
    EmbeddingsResponse:
      type: object
      additionalProperties: false
      required: [object, data, model, usage, x_router]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmbeddingRow"
        model:
          type: string
        usage:
          $ref: "#/components/schemas/EmbeddingsUsage"
        x_router:
          $ref: "#/components/schemas/EmbeddingsRouterMeta"
    EmbeddingsError:
      type: object
      additionalProperties: true
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, UNSUPPORTED_MODEL, IDEMPOTENCY_KEY_CONFLICT]
            message:
              type: string
            request_id:
              type: string
...
    ChatCompletionsRequest:
      type: object
      additionalProperties: false
      required: [model, messages]
      properties:
        model:
          type: string
          minLength: 1
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
        temperature:
          type: number
          minimum: 0
          maximum: 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
        max_tokens:
          type: integer
          minimum: 1
...
    RoutingExplainRequest:
      type: object
      additionalProperties: false
      required: [model]
      properties:
        model:
          type: string
          minLength: 1
        routing_preset:
          type: string
          enum: [cost, latency, success, balanced]
        region_preference:
          type: string
          enum: [US, EU, APAC]
...